Engine: Changed the text handling to use UTF-8 internally. Strings in games made earlier than 2.2 (the upcoming version) or earlier are converted from Windows-1252 when loaded.
Warning! This is a work-in-progress. Things are a bit broken right now.

Not done yet: Make sure that the msgBox and msgBoxQuestion functions display their strings properly on Windows and Linux now that they're fed Unicode.

Also not done: Update the Dev Kit to use Unicode, as that's what the engine now expects from new games. Otherwise the engine will crash.

git-svn-id: svn://svn.code.sf.net/p/opensludge/code@512 27a91dd4-ec73-4ff0-838b-64dfa05f697e