Added function freeSpriteBanks(spriteBank)
It frees the memory used by spriteBank (the argument should be a file handle), unless that sprite bank is currently being used, in which case it does nothing. The function returns TRUE on success and FALSE if the sprite bank can't be deleted (either because it isn't loaded, or because it is in use).

git-svn-id: svn://svn.code.sf.net/p/opensludge/code@642 27a91dd4-ec73-4ff0-838b-64dfa05f697e