Fix alpha blending of characters during freeze
The blending function did not set a correct value for resulting
alpha channel. This had no effect on resulting image when drawing
directly onto the screen, as the alpha channel was completely
ignored in that case - it did however influence the resulting image
if it was rendered first into a backdrop and only then displayed
on screen. Incorrect alpha value would then cause unwanted blending
to happen while drawing the backdrop, causing annoying darkening
of sprites around their edges.